Move wintun configuration into separate mod
